Endpoint to test push status
