+++ controllers/MobileAppController.php
@@ -6,18 +6,11 @@ namespace humhub\modules\fcmPush\controllers;
 use humhub\modules\admin\components\Controller;
 use humhub\modules\admin\notifications\NewVersionAvailable;
 use humhub\modules\fcmPush\models\FcmUser;
-use humhub\modules\fcmPush\Module;
 use humhub\modules\user\models\User;
 use Yii;
-use yii\web\Response;
 
 class MobileAppController extends Controller
 {
-    /**
-     * @inheritdoc
-     * @var Module $module
-     */
-    public $module;
 
     /**
      * Renders the index view for the module
@@ -50,29 +43,6 @@ class MobileAppController extends Controller
         }
 
         return $this->render('index');
-    }
-
-    public function actionStatus()
-    {
-        if (!$this->module->isActivated) {
-            return $this->returnStatus(404,'FCM Module is not installed');
-        }
-
-        if (!$this->module->getDriverService()->hasConfiguredDriver()) {
-            return $this->returnStatus(501,'Push Proxy is not configured');
-        }
-
-        return $this->returnStatus(200,'OK, Push (Proxy) is correctly configured');
-    }
-
-    private function returnStatus(int $code, string $message): Response
-    {
-        Yii::$app->response->statusCode = $code;
-
-        return $this->asJson([
-            'code' => $code,
-            'message' => $message,
-        ]);
     }
 
 }
